requirements-simple.txt
# Core Data Processing (without PySpark)
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# Database
psycopg2-binary==2.9.7
pymongo==4.4.1
sqlalchemy==2.0.19

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6

# Natural Language Processing
nltk==3.8.1
textblob==0.17.1

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web Framework & API
fastapi==0.101.1
uvicorn==0.23.2
streamlit==1.25.0

# Time Series Analysis
prophet==1.1.4
statsmodels==0.14.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.65.0

# Development Tools
jupyter==1.0.0
jupyterlab==4.0.7
ipykernel==6.25.0

# Docker & Deployment
gunicorn==21.2.0
